Added mono font for coding
After doing so many research, I found this font to be the best for coding. It is a mono font and it is very easy to read. I have added the font in the fonts folder. The font is ComicShannsMono. It has 3 variants, Regular, Bold and Propo. I have added all the variants in the fonts folder. I have also added the license and readme file for the font. I have also added the font in the .gitignore file so that it is not uploaded to the repository.